+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   JSON, not `changeowner`. This is a breaking change for the clients.
 - Public interface between contract and runtime no longer uses `String` to
   represent an error, but rather serializes `ApiError` as a rich JSON error.
+- Return value from `env.write_db` and `env.remove_db` to allow error reporting.
 
 **cosmwasm**